requirements.dev.txt
# Debugging
ipdb

# Coding standard
flake8
pep8-naming

# Pytest framework
pytest===2.8.4
pytest-cov
pytest-logging

# Documentation
sphinx
sphinx_rtd_theme
sphinxcontrib-plantuml
autoapi

# Topology libraries
topology_lib_ip
topology_lib_ping

# Remove when Topology 1.8 is released
-e git+git@github.com:matthew-m-hall/topology.git@master#egg=topology
